Separation of validation & routing
